scripts now generate gitkeeps
The scripts already generated the directories, which is enough to
satisfy jekyll when running locally. However, if the directories end up
empty during script generation, then there will be nothing to commit
back to git, and git will ignore/omit the directory. If the directory
doesn't exist in the git repo, then the GitHub Pages build fails.

Further, simply having the .gitkeep file in the repo doesn't help
because the scripts remove the directories entirely before proceeding.
So it's more robust for the scripts (which already remove and recreate
the directories) to ensure the gitkeeps are created themselves.