Replace phony 'yesterday' task with synthetic API file tasks
Instead of creating a dummy 'yesterday' task which forces re-building of
dependants when the json files are out of date, we now define a rule for
synthesizing tasks matching the API route for the analytics files.
This serves two purposes:
1. It does the filename-to-apiurl mapping in the "sources" side of the
   rule hash. This is the expected convention anyway for filenames,
   now we're just treating URLs as "remote files".
2. The synthetic tasks for the api routes are created as instances of
   the new HttpResourceTask. This new task class exposes HTTP Resources
   to rake in a similar fashion as FileTasks. In particular, the
   timestamp for the resource corresponds to its Last-Modified HTTP
   header. (With a convenience for reading the resource body.)

This is more semantically correct as the api itself is now treated very
much similarly to a file on disk. It is consulted for its timestamp when
determining if its dependants must be rebuilt. And there is a direct mapping
between the api "filename" (source) and the output filename (target).
This is what make/rake was built for!