Fix state issue with HttpResourceTask
Apparently, OpenURI#read can only be called once.
The value is not buffered or preserved. Subsequent reads return the
empty string. Somewhere, read was getting called a second time, so the
very first http task was returning "" for the contents.

This change now memoizes the results of read and json so the body is
not lost (and the JSON isn't parsed more often than necessary). This
introduces a tradeoff between memory consumption and performance, since
the json data for all files will be saved in memory. Memory-sensitive
scenarios will want to override this.

Also, improve the filter/map to use grep which filters and maps at the
same time. (Needs only a single pass of the array.)